shape: ray, sphere, and bbx
Add basic component Shape for different types of shapes which
provides the vertex, index, model transform, and calculation of
ray intersection. Extending Shape to create Sphere and Cube.

Add axis aligned bounding box (AABB) components which also help
estimate the ray intersection of Cube. Interval component is used
to build AABB.

Add an demo example which is used to test the intersection of view
ray (from camera to cursor) with the sphere and cube objects.