bvh: equal count and SAH based bvhs
Implement the equal count and SAH based BVHs. To facilitate the
SAH, we add a new method called ray_intersect_cost to
RayIntersectObject. We also modified partition method to allow
a customized predict method. In the demo, we can use LEFT and
RIGHT to change the BVH type and the ray intersect cost of the
BVH will be output. The BVH's ray intersect cost is estimated
with SAH, so in most cases, the SAH based BVH should have the
lowest cost.